+++ benchmark/federation/services/inventory/index.js
@@ -1,5 +1,5 @@
 const { gql } = require('graphql-tag');
-const { buildFederatedSchema } = require('@apollo/federation');
+const { buildSubgraphSchema } = require('@apollo/subgraph');
 
 const typeDefs = gql`
   extend type Product @key(fields: "upc") {
@@ -28,7 +28,7 @@ const resolvers = {
   },
 };
 
-const schema = buildFederatedSchema([
+const schema = buildSubgraphSchema([
   {
     typeDefs,
     resolvers,