+++ packages/delegate/src/delegateToSchema.ts
@@ -16,7 +16,6 @@ import {
   Executor,
   ExecutionRequest,
   Maybe,
-  AggregateError,
   isAsyncIterable,
   getDefinedRootType,
   memoize1,
@@ -83,7 +82,11 @@ function getDelegationReturnType(
   fieldName: string
 ): GraphQLOutputType {
   const rootType = getDefinedRootType(targetSchema, operation);
-  return rootType.getFields()[fieldName].type;
+  const rootFieldType = rootType.getFields()[fieldName];
+  if (!rootFieldType) {
+    throw new Error(`Unable to find field '${fieldName}' in type '${rootType}'.`);
+  }
+  return rootFieldType.type;
 }
 
 export function delegateRequest<