+++ packages/federation/test/fixtures/gateway/inventory/index.ts
@@ -1,0 +1,38 @@
+import { IResolvers } from '@graphql-tools/utils';
+import { inspect } from 'util';
+
+export const typeDefs = /* GraphQL */ `
+  type Product @key(fields: "upc") @extends {
+    upc: String! @external
+    weight: Int @external
+    price: Int @external
+    inStock: Boolean
+    shippingEstimate: Int @requires(fields: "price weight")
+  }
+`;
+
+export const resolvers: IResolvers = {
+  Product: {
+    __resolveReference(object) {
+      return {
+        ...object,
+        ...inventory.find(product => product.upc === object.upc),
+      };
+    },
+    shippingEstimate(object) {
+      if (object.price == null || object.weight == null) {
+        throw new Error(`${inspect(object)} doesn't have required fields; "price" and "weight".`);
+      }
+      // free for expensive items
+      if (object.price > 1000) return 0;
+      // estimate is based on weight
+      return object.weight * 0.5;
+    },
+  },
+};
+
+const inventory = [
+  { upc: '1', inStock: true },
+  { upc: '2', inStock: false },
+  { upc: '3', inStock: true },
+];