+++ packages/merge/src/typedefs-mergers/merge-typedefs.ts
@@ -8,9 +8,10 @@ import {
   OperationTypeNode,
   isDefinitionNode,
   ParseOptions,
+  DirectiveDefinitionNode,
 } from 'graphql';
 import { CompareFn, defaultStringComparator, isSourceTypes, isStringTypes } from './utils.js';
-import { MergedResultMap, mergeGraphQLNodes, schemaDefSymbol } from './merge-nodes.js';
+import { mergeGraphQLNodes, schemaDefSymbol } from './merge-nodes.js';
 import {
   getDocumentNodeFromSchema,
   GetDocumentNodeFromSchemaOptions,
@@ -135,40 +136,65 @@ export function mergeTypeDefs(typeSource: TypeSource, config?: Partial<Config>):
 function visitTypeSources(
   typeSource: TypeSource,
   options: ParseOptions & GetDocumentNodeFromSchemaOptions,
+  allDirectives: DirectiveDefinitionNode[] = [],
   allNodes: DefinitionNode[] = [],
   visitedTypeSources = new Set<TypeSource>()
 ) {
   if (typeSource && !visitedTypeSources.has(typeSource)) {
     visitedTypeSources.add(typeSource);
     if (typeof typeSource === 'function') {
-      visitTypeSources(typeSource(), options, allNodes, visitedTypeSources);
+      visitTypeSources(typeSource(), options, allDirectives, allNodes, visitedTypeSources);
     } else if (Array.isArray(typeSource)) {
       for (const type of typeSource) {
-        visitTypeSources(type, options, allNodes, visitedTypeSources);
+        visitTypeSources(type, options, allDirectives, allNodes, visitedTypeSources);
       }
     } else if (isSchema(typeSource)) {
       const documentNode = getDocumentNodeFromSchema(typeSource, options);
-      visitTypeSources(documentNode.definitions as DefinitionNode[], options, allNodes, visitedTypeSources);
+      visitTypeSources(
+        documentNode.definitions as DefinitionNode[],
+        options,
+        allDirectives,
+        allNodes,
+        visitedTypeSources
+      );
     } else if (isStringTypes(typeSource) || isSourceTypes(typeSource)) {
       const documentNode = parse(typeSource, options);
-      visitTypeSources(documentNode.definitions as DefinitionNode[], options, allNodes, visitedTypeSources);
+      visitTypeSources(
+        documentNode.definitions as DefinitionNode[],
+        options,
+        allDirectives,
+        allNodes,
+        visitedTypeSources
+      );
     } else if (typeof typeSource === 'object' && isDefinitionNode(typeSource)) {
-      allNodes.push(typeSource);
+      if (typeSource.kind === Kind.DIRECTIVE_DEFINITION) {
+        allDirectives.push(typeSource);
+      } else {
+        allNodes.push(typeSource);
+      }
     } else if (isDocumentNode(typeSource)) {
-      visitTypeSources(typeSource.definitions as DefinitionNode[], options, allNodes, visitedTypeSources);
+      visitTypeSources(
+        typeSource.definitions as DefinitionNode[],
+        options,
+        allDirectives,
+        allNodes,
+        visitedTypeSources
+      );
     } else {
       throw new Error(`typeDefs must contain only strings, documents, schemas, or functions, got ${typeof typeSource}`);
     }
   }
-  return allNodes;
+  return { allDirectives, allNodes };
 }
 
 export function mergeGraphQLTypes(typeSource: TypeSource, config: Config): DefinitionNode[] {
   resetComments();
 
-  const allNodes = visitTypeSources(typeSource, config);
+  const { allDirectives, allNodes } = visitTypeSources(typeSource, config);
 
-  const mergedNodes: MergedResultMap = mergeGraphQLNodes(allNodes, config);
+  const mergedDirectives = mergeGraphQLNodes(allDirectives, config) as Record<string, DirectiveDefinitionNode>;
+
+  const mergedNodes = mergeGraphQLNodes(allNodes, config, mergedDirectives);
 
   if (config?.useSchemaDefinition) {
     // XXX: right now we don't handle multiple schema definitions