+++ packages/loaders/prisma/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @graphql-tools/prisma-loader
 
+## 8.0.0
+
+### Major Changes
+
+- [#5274](https://github.com/ardatan/graphql-tools/pull/5274) [`944a68e8`](https://github.com/ardatan/graphql-tools/commit/944a68e8becf9c86b4c97fd17c372d98a285b955) Thanks [@ardatan](https://github.com/ardatan)! - Drop Node 14 support. Require Node.js `>= 16`
+
+### Patch Changes
+
+- [#5274](https://github.com/ardatan/graphql-tools/pull/5274) [`944a68e8`](https://github.com/ardatan/graphql-tools/commit/944a68e8becf9c86b4c97fd17c372d98a285b955) Thanks [@ardatan](https://github.com/ardatan)! - dependencies updates:
+  - Updated dependency [`@whatwg-node/fetch@^0.9.0` ↗︎](https://www.npmjs.com/package/@whatwg-node/fetch/v/0.9.0) (from `^0.8.2`, in `dependencies`)
+- Updated dependencies [[`944a68e8`](https://github.com/ardatan/graphql-tools/commit/944a68e8becf9c86b4c97fd17c372d98a285b955), [`944a68e8`](https://github.com/ardatan/graphql-tools/commit/944a68e8becf9c86b4c97fd17c372d98a285b955), [`944a68e8`](https://github.com/ardatan/graphql-tools/commit/944a68e8becf9c86b4c97fd17c372d98a285b955)]:
+  - @graphql-tools/url-loader@8.0.0
+  - @graphql-tools/utils@10.0.0
+
+## 7.2.72
+
+### Patch Changes
+
+- [#5251](https://github.com/ardatan/graphql-tools/pull/5251) [`58a18e4d`](https://github.com/ardatan/graphql-tools/commit/58a18e4d60dfb43dff2ebdeca2865da0bbaa1c16) Thanks [@renovate](https://github.com/apps/renovate)! - dependencies updates:
+
+  - Updated dependency [`http-proxy-agent@^6.0.0` ↗︎](https://www.npmjs.com/package/http-proxy-agent/v/6.0.0) (from `^5.0.0`, in `dependencies`)
+
+- [#5252](https://github.com/ardatan/graphql-tools/pull/5252) [`8f79ded2`](https://github.com/ardatan/graphql-tools/commit/8f79ded25b82f68b8950ef326c90d4414f780912) Thanks [@renovate](https://github.com/apps/renovate)! - dependencies updates:
+  - Updated dependency [`https-proxy-agent@^6.0.0` ↗︎](https://www.npmjs.com/package/https-proxy-agent/v/6.0.0) (from `^5.0.0`, in `dependencies`)
+
+## 7.2.71
+
+### Patch Changes
+
+- [`bba3d32f`](https://github.com/ardatan/graphql-tools/commit/bba3d32f6c60c32530ba2e48aadb9baaf985978c) Thanks [@ardatan](https://github.com/ardatan)! - Unpin url loader
+
+## 7.2.70
+
+### Patch Changes
+
+- [#5166](https://github.com/ardatan/graphql-tools/pull/5166) [`807c9b54`](https://github.com/ardatan/graphql-tools/commit/807c9b547369903575c1182f891e9a89aadbaeb8) Thanks [@renovate](https://github.com/apps/renovate)! - dependencies updates:
+
+  - Updated dependency [`graphql-request@^6.0.0` ↗︎](https://www.npmjs.com/package/graphql-request/v/6.0.0) (from `^5.0.0`, in `dependencies`)
+
+- [`1c95368a`](https://github.com/ardatan/graphql-tools/commit/1c95368aea868be537d956ba5e994cde58dfee41) Thanks [@ardatan](https://github.com/ardatan)! - Use ranged versions for dependencies
+
+- Updated dependencies []:
+  - @graphql-tools/url-loader@7.17.18
+
 ## 7.2.69
 
 ### Patch Changes