+++ packages/optimize/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphql-tools/optimize
 
+## 2.0.0
+
+### Major Changes
+
+- [#5274](https://github.com/ardatan/graphql-tools/pull/5274) [`944a68e8`](https://github.com/ardatan/graphql-tools/commit/944a68e8becf9c86b4c97fd17c372d98a285b955) Thanks [@ardatan](https://github.com/ardatan)! - Drop Node 14 support. Require Node.js `>= 16`
+
+## 1.4.0
+
+### Minor Changes
+
+- [#5177](https://github.com/ardatan/graphql-tools/pull/5177) [`c6817c59`](https://github.com/ardatan/graphql-tools/commit/c6817c596314795165f531feca408f78d29e9f6d) Thanks [@AaronMoat](https://github.com/AaronMoat)! - feat(optimize): update optimize to also remove descriptions from directives
+
 ## 1.3.1
 
 ### Patch Changes