+++ packages/webpack-loader/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @graphql-tools/webpack-loader
 
+## 7.0.0
+
+### Major Changes
+
+- [#5274](https://github.com/ardatan/graphql-tools/pull/5274) [`944a68e8`](https://github.com/ardatan/graphql-tools/commit/944a68e8becf9c86b4c97fd17c372d98a285b955) Thanks [@ardatan](https://github.com/ardatan)! - Drop Node 14 support. Require Node.js `>= 16`
+
+### Patch Changes
+
+- Updated dependencies [[`944a68e8`](https://github.com/ardatan/graphql-tools/commit/944a68e8becf9c86b4c97fd17c372d98a285b955)]:
+  - @graphql-tools/optimize@2.0.0
+  - @graphql-tools/webpack-loader-runtime@7.0.0
+
+## 6.7.2
+
+### Patch Changes
+
+- Updated dependencies [[`c6817c59`](https://github.com/ardatan/graphql-tools/commit/c6817c596314795165f531feca408f78d29e9f6d)]:
+  - @graphql-tools/optimize@1.4.0
+
 ## 6.7.1
 
 ### Patch Changes